layout for bulk group
